# java-openapi-controller/src/main/java/org/org/numb/openapi/generator/gen/delegate/CurlApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.org.numb.openapi.generator.gen.delegate;

import org.org.numb.openapi.generator.gen.model.SomeObj;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-25T08:13:05.193599400+08:00[Asia/Shanghai]", comments = "Generator version: 7.14.0")
@Validated
@Tag(name = "curl", description = "the curl API")
public interface CurlApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /curl
     *
     * @param petId ID of pet that needs to be updated (required)
     * @param name Updated name of the pet (optional)
     * @param status Updated status of the pet (optional)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getCurl",
        tags = { "curl" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SomeObj.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/curl",
        produces = { "application/json" },
        consumes = { "application/x-www-form-urlencoded" }
    )
    
    default ResponseEntity<SomeObj> _getCurl(
        @NotNull @Parameter(name = "petId", description = "ID of pet that needs to be updated", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "petId", required = true) Long petId,
        @Parameter(name = "name", description = "Updated name of the pet") @Valid @RequestParam(value = "name", required = false) String name,
        @Parameter(name = "status", description = "Updated status of the pet") @Valid @RequestParam(value = "status", required = false) String status
    ) {
        return getCurl(petId, name, status);
    }

    // Override this method
    default  ResponseEntity<SomeObj> getCurl(Long petId, String name, String status) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"name\" : \"name\", \"active\" : true, \"$_type\" : \"SomeObjIdentifier\", \"id\" : 0, \"type\" : \"type\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /curl
     *
     * @param someObj  (optional)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "postCurl",
        tags = { "ping" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = SomeObj.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/curl",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<SomeObj> _postCurl(
        @Parameter(name = "SomeObj", description = "") @Valid @RequestBody(required = false) @Nullable SomeObj someObj
    ) {
        return postCurl(someObj);
    }

    // Override this method
    default  ResponseEntity<SomeObj> postCurl(SomeObj someObj) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"name\" : \"name\", \"active\" : true, \"$_type\" : \"SomeObjIdentifier\", \"id\" : 0, \"type\" : \"type\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}