# java-openapi-controller/src/main/java/org/org/numb/openapi/generator/gen/model/SimpleOneOf.java
package org.org.numb.openapi.generator.gen.model;

import java.net.URI;
import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonSubTypes;
import com.fasterxml.jackson.annotation.JsonTypeInfo;
import org.openapitools.jackson.nullable.JsonNullable;
import java.time.OffsetDateTime;
import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;


import java.util.*;
import jakarta.annotation.Generated;


@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-25T08:13:05.193599400+08:00[Asia/Shanghai]", comments = "Generator version: 7.14.0")
public interface SimpleOneOf {
}