# java-openapi-controller/src/main/resources/openapi/v1/openapi.yaml
openapi: 3.0.1
info:
  title: ping some object
  version: "1.0"
servers:
  - url: http://localhost:8082/
tags:
  - ping
  - curl
paths:
  /ping:
    get:
      operationId: getPing
      parameters:
        - description: ID of pet that needs to be updated
          explode: true
          in: query
          name: petId
          required: true
          schema:
            format: int64
            type: integer
          style: form
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/getPing_request"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SomeObj"
          description: OK
      tags:
        - ping
      x-streaming: true
      x-group-parameters: true
      x-content-type: application/x-www-form-urlencoded
      x-accepts:
        - application/json
    post:
      operationId: postPing
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SomeObj"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SomeObj"
          description: OK
      tags:
        - ping
      x-streaming: true
      x-content-type: application/json
      x-accepts:
        - application/json
  /curl:
    get:
      operationId: getCurl
      parameters:
        - description: ID of pet that needs to be updated
          explode: true
          in: query
          name: petId
          required: true
          schema:
            format: int64
            type: integer
          style: form
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/getCurl_request"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SomeObj"
          description: OK
      tags:
        - curl
      x-streaming: true
      x-group-parameters: true
      x-content-type: application/x-www-form-urlencoded
      x-accepts:
        - application/json
    post:
      operationId: postCurl
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SomeObj"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SomeObj"
          description: OK
      tags:
        - ping
      x-streaming: true
      x-content-type: application/json
      x-accepts:
        - application/json
components:
  schemas:
    SomeObj:
      example:
        name: name
        active: true
        $_type: SomeObjIdentifier
        id: 0
        type: type
      properties:
        $_type:
          default: SomeObjIdentifier
          enum:
            - SomeObjIdentifier
          type: string
        id:
          format: int64
          type: integer
        name:
          type: string
        active:
          type: boolean
        type:
          type: string
      type: object
    SimpleOneOf:
      oneOf:
        - type: string
        - type: integer
    getPing_request:
      properties:
        name:
          description: Updated name of the pet
          type: string
        status:
          description: Updated status of the pet
          type: string
      type: object
    getCurl_request:
      properties:
        name:
          description: Updated name of the pet
          type: string
        status:
          description: Updated status of the pet
          type: string
      type: object